+++ tests/TestBayesModels.cc
@@ -2,9 +2,6 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/catch_approx.hpp>
 #include <catch2/generators/catch_generators.hpp>
-#include <vector>
-#include <map>
-#include <string>
 #include "KDB.h"
 #include "TAN.h"
 #include "SPODE.h"
@@ -16,12 +13,9 @@
 #include "AODELd.h"
 #include "TestUtils.h"
 
-TEST_CASE("Library check version", "[BayesNet]")
-{
-    auto clf = bayesnet::KDB(2);
-    REQUIRE(clf.getVersion() == "1.0.2");
-}
-TEST_CASE("Test Bayesian Classifiers score", "[BayesNet]")
+const std::string ACTUAL_VERSION = "1.0.3";
+
+TEST_CASE("Test Bayesian Classifiers score & version", "[BayesNet]")
 {
     map <pair<std::string, std::string>, float> scores = {
         // Diabetes
@@ -37,87 +31,34 @@ TEST_CASE("Test Bayesian Classifiers score", "[BayesNet]")
         {{"iris", "AODE"}, 0.973333}, {{"iris", "KDB"}, 0.973333}, {{"iris", "SPODE"}, 0.973333}, {{"iris", "TAN"}, 0.973333},
         {{"iris", "AODELd"}, 0.973333}, {{"iris", "KDBLd"}, 0.973333}, {{"iris", "SPODELd"}, 0.96f}, {{"iris", "TANLd"}, 0.97333f}, {{"iris", "BoostAODE"}, 0.98f}
     };
+    std::map<std::string, bayesnet::BaseClassifier*> models = {
+        {"AODE", new bayesnet::AODE()}, {"AODELd", new bayesnet::AODELd()},
+        {"BoostAODE", new bayesnet::BoostAODE()},
+        {"KDB", new bayesnet::KDB(2)}, {"KDBLd", new bayesnet::KDBLd(2)},
+        {"SPODE", new bayesnet::SPODE(1)}, {"SPODELd", new bayesnet::SPODELd(1)},
+        {"TAN", new bayesnet::TAN()}, {"TANLd", new bayesnet::TANLd()}
+    };
+    std::string name = GENERATE("AODE", "AODELd", "KDB", "KDBLd", "SPODE", "SPODELd", "TAN", "TANLd");
+    auto clf = models[name];
 
-    std::string file_name = GENERATE("glass", "iris", "ecoli", "diabetes");
-    auto raw = RawDatasets(file_name, false);
-
-    SECTION("Test TAN classifier (" + file_name + ")")
+    SECTION("Test " + name + " classifier")
     {
-        auto clf = bayesnet::TAN();
-        clf.fit(raw.Xv, raw.yv, raw.featuresv, raw.classNamev, raw.statesv);
-        auto score = clf.score(raw.Xv, raw.yv);
-        //scores[{file_name, "TAN"}] = score;
-        REQUIRE(score == Catch::Approx(scores[{file_name, "TAN"}]).epsilon(raw.epsilon));
+        for (const std::string& file_name : { "glass", "iris", "ecoli", "diabetes" }) {
+            auto clf = models[name];
+            auto discretize = name.substr(name.length() - 2) != "Ld";
+            auto raw = RawDatasets(file_name, discretize);
+            clf->fit(raw.Xt, raw.yt, raw.featurest, raw.classNamet, raw.statest);
+            auto score = clf->score(raw.Xt, raw.yt);
+            INFO("File: " + file_name);
+            REQUIRE(score == Catch::Approx(scores[{file_name, name}]).epsilon(raw.epsilon));
+        }
     }
-    SECTION("Test TANLd classifier (" + file_name + ")")
+    SECTION("Library check version")
     {
-        auto clf = bayesnet::TANLd();
-        clf.fit(raw.Xt, raw.yt, raw.featurest, raw.classNamet, raw.statest);
-        auto score = clf.score(raw.Xt, raw.yt);
-        //scores[{file_name, "TANLd"}] = score;
-        REQUIRE(score == Catch::Approx(scores[{file_name, "TANLd"}]).epsilon(raw.epsilon));
+        INFO("Checking version of " + name + " classifier");
+        REQUIRE(clf->getVersion() == ACTUAL_VERSION);
     }
-    SECTION("Test KDB classifier (" + file_name + ")")
-    {
-        auto clf = bayesnet::KDB(2);
-        clf.fit(raw.Xv, raw.yv, raw.featuresv, raw.classNamev, raw.statesv);
-        auto score = clf.score(raw.Xv, raw.yv);
-        //scores[{file_name, "KDB"}] = score;
-        REQUIRE(score == Catch::Approx(scores[{file_name, "KDB"
-        }]).epsilon(raw.epsilon));
-    }
-    SECTION("Test KDBLd classifier (" + file_name + ")")
-    {
-        auto clf = bayesnet::KDBLd(2);
-        clf.fit(raw.Xt, raw.yt, raw.featurest, raw.classNamet, raw.statest);
-        auto score = clf.score(raw.Xt, raw.yt);
-        //scores[{file_name, "KDBLd"}] = score;
-        REQUIRE(score == Catch::Approx(scores[{file_name, "KDBLd"
-        }]).epsilon(raw.epsilon));
-    }
-    SECTION("Test SPODE classifier (" + file_name + ")")
-    {
-        auto clf = bayesnet::SPODE(1);
-        clf.fit(raw.Xv, raw.yv, raw.featuresv, raw.classNamev, raw.statesv);
-        auto score = clf.score(raw.Xv, raw.yv);
-        // scores[{file_name, "SPODE"}] = score;
-        REQUIRE(score == Catch::Approx(scores[{file_name, "SPODE"}]).epsilon(raw.epsilon));
-    }
-    SECTION("Test SPODELd classifier (" + file_name + ")")
-    {
-        auto clf = bayesnet::SPODELd(1);
-        clf.fit(raw.Xt, raw.yt, raw.featurest, raw.classNamet, raw.statest);
-        auto score = clf.score(raw.Xt, raw.yt);
-        // scores[{file_name, "SPODELd"}] = score;
-        REQUIRE(score == Catch::Approx(scores[{file_name, "SPODELd"}]).epsilon(raw.epsilon));
-    }
-    SECTION("Test AODE classifier (" + file_name + ")")
-    {
-        auto clf = bayesnet::AODE();
-        clf.fit(raw.Xv, raw.yv, raw.featuresv, raw.classNamev, raw.statesv);
-        auto score = clf.score(raw.Xv, raw.yv);
-        // scores[{file_name, "AODE"}] = score;
-        REQUIRE(score == Catch::Approx(scores[{file_name, "AODE"}]).epsilon(raw.epsilon));
-    }
-    SECTION("Test AODELd classifier (" + file_name + ")")
-    {
-        auto clf = bayesnet::AODELd();
-        clf.fit(raw.Xt, raw.yt, raw.featurest, raw.classNamet, raw.statest);
-        auto score = clf.score(raw.Xt, raw.yt);
-        // scores[{file_name, "AODELd"}] = score;
-        REQUIRE(score == Catch::Approx(scores[{file_name, "AODELd"}]).epsilon(raw.epsilon));
-    }
-    SECTION("Test BoostAODE classifier (" + file_name + ")")
-    {
-        auto clf = bayesnet::BoostAODE();
-        clf.fit(raw.Xv, raw.yv, raw.featuresv, raw.classNamev, raw.statesv);
-        auto score = clf.score(raw.Xv, raw.yv);
-        // scores[{file_name, "BoostAODE"}] = score;
-        REQUIRE(score == Catch::Approx(scores[{file_name, "BoostAODE"}]).epsilon(raw.epsilon));
-    }
-    // for (auto scores : scores) {
-    //     std::cout << "{{\"" << scores.first.first << "\", \"" << scores.first.second << "\"}, " << scores.second << "}, ";
-    // }
+    delete clf;
 }
 TEST_CASE("Models features", "[BayesNet]")
 {
@@ -133,6 +74,8 @@ TEST_CASE("Models features", "[BayesNet]")
     clf.fit(raw.Xv, raw.yv, raw.featuresv, raw.classNamev, raw.statesv);
     REQUIRE(clf.getNumberOfNodes() == 5);
     REQUIRE(clf.getNumberOfEdges() == 7);
+    REQUIRE(clf.getNumberOfStates() == 19);
+    REQUIRE(clf.getClassNumStates() == 3);
     REQUIRE(clf.show() == std::vector<std::string>{"class -> sepallength, sepalwidth, petallength, petalwidth, ", "petallength -> sepallength, ", "petalwidth -> ", "sepallength -> sepalwidth, ", "sepalwidth -> petalwidth, "});
     REQUIRE(clf.graph("Test") == graph);
 }
@@ -156,16 +99,15 @@ TEST_CASE("BoostAODE feature_select CFS", "[BayesNet]")
     REQUIRE(clf.getNotes()[0] == "Used features in initialization: 6 of 9 with CFS");
     REQUIRE(clf.getNotes()[1] == "Number of models: 9");
 }
-TEST_CASE("BoostAODE test used features in train note", "[BayesNet]")
+TEST_CASE("BoostAODE test used features in train note and score", "[BayesNet]")
 {
     auto raw = RawDatasets("diabetes", true);
-    auto clf = bayesnet::BoostAODE();
+    auto clf = bayesnet::BoostAODE(true);
     clf.setHyperparameters({
         {"ascending",true},
         {"convergence", true},
         {"repeatSparent",true},
         {"select_features","CFS"},
-        {"tolerance", 3}
         });
     clf.fit(raw.Xv, raw.yv, raw.featuresv, raw.classNamev, raw.statesv);
     REQUIRE(clf.getNumberOfNodes() == 72);
@@ -174,4 +116,109 @@ TEST_CASE("BoostAODE test used features in train note", "[BayesNet]")
     REQUIRE(clf.getNotes()[0] == "Used features in initialization: 6 of 8 with CFS");
     REQUIRE(clf.getNotes()[1] == "Used features in train: 7 of 8");
     REQUIRE(clf.getNotes()[2] == "Number of models: 8");
+    auto score = clf.score(raw.Xv, raw.yv);
+    auto scoret = clf.score(raw.Xt, raw.yt);
+    REQUIRE(score == Catch::Approx(0.8138).epsilon(raw.epsilon));
+    REQUIRE(scoret == Catch::Approx(0.8138).epsilon(raw.epsilon));
+}
+TEST_CASE("Model predict_proba", "[BayesNet]")
+{
+    std::string model = GENERATE("TAN", "SPODE", "BoostAODEproba", "BoostAODEvoting");
+    auto res_prob_tan = std::vector<std::vector<double>>({
+    { 0.00375671, 0.994457, 0.00178621 },
+    { 0.00137462, 0.992734, 0.00589123 },
+    { 0.00137462, 0.992734, 0.00589123 },
+    { 0.00137462, 0.992734, 0.00589123 },
+    { 0.00218225, 0.992877, 0.00494094 },
+    { 0.00494209, 0.0978534, 0.897205 },
+    { 0.0054192, 0.974275, 0.0203054 },
+    { 0.00433012, 0.985054, 0.0106159 },
+    { 0.000860806, 0.996922, 0.00221698 }
+        });
+    auto res_prob_spode = std::vector<std::vector<double>>({
+     {0.00419032, 0.994247, 0.00156265},
+     {0.00172808, 0.993433, 0.00483862},
+     {0.00172808, 0.993433, 0.00483862},
+     {0.00172808, 0.993433, 0.00483862},
+     {0.00279211, 0.993737, 0.00347077},
+     {0.0120674, 0.357909, 0.630024},
+     {0.00386239, 0.913919, 0.0822185},
+     {0.0244389, 0.966447, 0.00911374},
+     {0.003135, 0.991799, 0.0050661}
+        });
+    auto res_prob_baode = std::vector<std::vector<double>>({
+        {0.00803291, 0.9676, 0.0243672},
+        {0.00398714, 0.945126, 0.050887},
+        {0.00398714, 0.945126, 0.050887},
+        {0.00398714, 0.945126, 0.050887},
+        {0.00189227, 0.859575, 0.138533},
+        {0.0118341, 0.442149, 0.546017},
+        {0.0216135, 0.785781, 0.192605},
+        {0.0204803, 0.844276, 0.135244},
+        {0.00576313, 0.961665, 0.0325716},
+        });
+    auto res_prob_voting = std::vector<std::vector<double>>({
+        {0, 1, 0},
+        {0, 1, 0},
+        {0, 1, 0},
+        {0, 1, 0},
+        {0, 1, 0},
+        {0, 0.447909, 0.552091},
+        {0, 0.811482, 0.188517},
+        {0, 1, 0},
+        {0, 1, 0}
+        });
+    std::map<std::string, std::vector<std::vector<double>>> res_prob = { {"TAN", res_prob_tan}, {"SPODE", res_prob_spode} , {"BoostAODEproba", res_prob_baode }, {"BoostAODEvoting", res_prob_voting } };
+    std::map<std::string, bayesnet::BaseClassifier*> models = { {"TAN", new bayesnet::TAN()}, {"SPODE", new bayesnet::SPODE(0)}, {"BoostAODEproba", new bayesnet::BoostAODE(false)}, {"BoostAODEvoting", new bayesnet::BoostAODE(true)} };
+    int init_index = 78;
+    auto raw = RawDatasets("iris", true);
+
+    SECTION("Test " + model + " predict_proba")
+    {
+        auto clf = models[model];
+        clf->fit(raw.Xv, raw.yv, raw.featuresv, raw.classNamev, raw.statesv);
+        auto y_pred_proba = clf->predict_proba(raw.Xv);
+        auto y_pred = clf->predict(raw.Xv);
+        auto yt_pred = clf->predict(raw.Xt);
+        auto yt_pred_proba = clf->predict_proba(raw.Xt);
+        REQUIRE(y_pred.size() == yt_pred.size(0));
+        REQUIRE(y_pred.size() == y_pred_proba.size());
+        REQUIRE(y_pred.size() == yt_pred_proba.size(0));
+        REQUIRE(y_pred.size() == raw.yv.size());
+        REQUIRE(y_pred_proba[0].size() == 3);
+        REQUIRE(yt_pred_proba.size(1) == y_pred_proba[0].size());
+        for (int i = 0; i < y_pred_proba.size(); ++i) {
+            auto maxElem = max_element(y_pred_proba[i].begin(), y_pred_proba[i].end());
+            int predictedClass = distance(y_pred_proba[i].begin(), maxElem);
+            REQUIRE(predictedClass == y_pred[i]);
+            // Check predict is coherent with predict_proba
+            REQUIRE(yt_pred_proba[i].argmax().item<int>() == y_pred[i]);
+        }
+        // Check predict_proba values for vectors and tensors
+        for (int i = 0; i < res_prob.size(); i++) {
+            REQUIRE(y_pred[i] == yt_pred[i].item<int>());
+            for (int j = 0; j < 3; j++) {
+                REQUIRE(res_prob[model][i][j] == Catch::Approx(y_pred_proba[i + init_index][j]).epsilon(raw.epsilon));
+                REQUIRE(res_prob[model][i][j] == Catch::Approx(yt_pred_proba[i + init_index][j].item<double>()).epsilon(raw.epsilon));
+            }
+        }
+        delete clf;
+    }
+}
+TEST_CASE("BoostAODE voting-proba", "[BayesNet]")
+{
+    auto raw = RawDatasets("iris", false);
+    auto clf = bayesnet::BoostAODE(false);
+    clf.fit(raw.Xv, raw.yv, raw.featuresv, raw.classNamev, raw.statesv);
+    auto score_proba = clf.score(raw.Xv, raw.yv);
+    auto pred_proba = clf.predict_proba(raw.Xv);
+    clf.setHyperparameters({
+        {"predict_voting",true},
+        });
+    auto score_voting = clf.score(raw.Xv, raw.yv);
+    auto pred_voting = clf.predict_proba(raw.Xv);
+    REQUIRE(score_proba == Catch::Approx(0.97333).epsilon(raw.epsilon));
+    REQUIRE(score_voting == Catch::Approx(0.98).epsilon(raw.epsilon));
+    REQUIRE(pred_voting[83][2] == Catch::Approx(0.552091).epsilon(raw.epsilon));
+    REQUIRE(pred_proba[83][2] == Catch::Approx(0.546017).epsilon(raw.epsilon));
 }